# swagger.yaml
...
swagger: '2.0'
info:
  title: event-broker API documentation
  termsOfService: http://swagger.io/terms/
  description: ''
  version: '2017-08-10'
  contact:
    name: sh20xyqi
    email: sh20xyqi@studserv.uni-leipzig.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - http
host: fgseitsrancher.wifa.intern.uni-leipzig.de:8000
basePath: /event-broker
produces:
  - text/plain
...
paths:
...
  /pool/assignments/{poolid}:
    get:
      tags:
        - event-broker
      summary:
        Returns all PoolAssignments behind the passed Id.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: poolid
          in: path
          description: Idenifies the PoolId using the passed Id.
          required: true
          type: string
      responses:
        200:
          description: Returns all FilterPoolAssignments using the passed PoolId.
          schema:
            type: array
            items:
              $ref: '#/definitions/FilterPoolAssignment'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - Server was not able to return all PoolAssignments from database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /pool/command/{poolid}/{poolsize}:
    get:
      tags:
        - event-broker
      summary:
        Returns a FilterPoolCommand using the passed Pool Id and the passed pool size.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: poolid
          in: path
          description: Idenifies the PoolId using the passed Id.
          required: true
          type: string
        - name: poolsize
          in: path
          description: Idenifies the pool_size using the passed poolsize.
          required: true
          type: string
      responses:
        200:
          description: Returns the FilterPoolCommand using the passed PoolId and pool_size.
          schema:
            $ref: '#/definitions/FilterPoolCommand'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - Server was not able to return FilterPoolCommand from database.
          schema:
            $ref: '#/definitions/ErrorMessage'
...
definitions:
...
  FilterPoolCommand:
    type: object
    properties:
      Command:
        type: string
        format: string
      Assignment:
        type: array
        items:
          $ref: '#/definitions/FilterPoolAssignment'
    required:
      - Command
      - Assignment
  FilterPoolAssignment:
    type: object
    properties:
      rule_id:
        type: string
        format: string
      routing_info:
        type: string
        format: string
    required:
      - rule_id
      - routing_info
  ErrorMessage:
    type: object
    properties:
      status_code:
        type: integer
        format: int32
      message:
        type: string
        format: string
      error_code:
        type: string
        format: string
      detail:
        type: array
        items:
          type: string
    required:
      - status_code
      - message
      - error_code
      - detail